requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.45.0
anyio==4.8.0
async-timeout==4.0.3
attrs==24.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dataclasses-json==0.6.7
Deprecated==1.2.17
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
emoji==2.14.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
faiss-cpu==1.9.0.post1
fastapi==0.115.7
ffmpy==0.5.0
filelock==3.17.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
gradio==5.13.1
gradio_client==1.6.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
hyperframe==6.1.0
idna==3.10
img2pdf==0.5.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-text-splitters==0.3.5
langdetect==1.0.9
langsmith==0.3.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.23.0
ndjson==0.3.1
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
ocrmypdf==16.3.1
olefile==0.47
openai==0.28.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pdfminer.six==20240706
pi_heif==0.21.0
pikepdf==9.5.1
pillow==11.1.0
pluggy==1.5.0
portalocker==2.10.1
propcache==0.2.1
protobuf==5.29.3
psutil==5.9.8
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pymongo==4.10.1
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.13.2
RapidFuzz==3.11.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
ruff==0.9.3
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spaces==0.32.0
SQLAlchemy==2.0.37
starlette==0.45.3
streamlit==1.41.1
tenacity==9.0.0
tiktoken==0.8.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.16.15
unstructured-client==0.29.0
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
webencodings==0.5.1
websockets==14.2
wrapt==1.17.2
yarl==1.18.3
zstandard==0.23.0
